Instances, instance manager and versions
The `Instance` and `InstanceManager` classes have been added for storing data about SuperTux instances and managing them.

The `Version` class is to-be-used as a base for each SuperTux version present in the launcher. It contains virtual methods, allowing to incorporate version differences, so they can all be handled easily.
A class for the version 0.6.3 has been added for testing.

Along with these main classes, a lot of other helper enumerators and functions have also been added.