Use `InstanceManager` as a model for `InstanceList`
`InstanceManager` now also inherits `QStandardItemModel`, allowing it to be used as a model for `InstanceList`, which now inherits `QTreeView`.

The model/view framework is being used to prevent refreshing the `InstanceList` every time a change happens in the `InstanceManager`.

The listed instances are now also sorted in ascending order by their creation time.