Parse versions from files, `VersionManager`
The base for the previous source class interface `Version` system was replaced with one, where versions are loaded from data files. This allows for more flexibility in adding/modifying/deleting versions from the launcher.

A `VersionManager` was added, which creates all versions from their respective file, as well as provides some `get()` functions for one or all instances.

Only one "data/versions" subfolder's contents are copied to the binary "data/versions" folder, so only the version files for the target OS and architecture are available.