Re-use general options menu as instance options menu, view build/run logs
The general options menu, added in 0994b94bce60999b93dd51c00b51fdc3bfe9d42a, was re-used as an instance options menu, because it's more practical to implement a way to view per-build/run instance logs, rather than a full log of the launcher's output.

The menu will be re-introduced in the future when its time comes.