Manage instance processes using `QProcess`
`QProcess` is now used for managing and launching instance processes.

Versions no longer have a "run-format" entry, but rather a "support" one, which outlines what features the version supports. That is used for determining the command line arguments to be passed to processes of that version.

TODO: Deny removing or launching another processes of an already running instance.